Converted organism stories into individual stories.
