+++ stories/organisms/default.stories.tsx
@@ -1,0 +1,9 @@
+import Default from "../../components/organisms/Default/default";
+
+const storyConfig = {
+  title: "Design System/Organisms/Default"
+};
+
+export default storyConfig;
+
+export const DefaultOrganism = () => <Default />;
