+++ stories/organisms/footer.stories.tsx
@@ -1,0 +1,9 @@
+import Footer from "../../components/organisms/footer/footer";
+
+const storyConfig = {
+  title: "Design System/Organisms/Footer"
+};
+
+export default storyConfig;
+
+export const FooterOrganism = () => <Footer />;
