+++ stories/organisms/header.stories.tsx
@@ -1,0 +1,9 @@
+import Header from "../../components/organisms/header/header";
+
+const storyConfig = {
+  title: "Design System/Organisms/Header"
+};
+
+export default storyConfig;
+
+export const HeaderOrganism = () => <Header />;
