+++ stories/organisms/tool-nav.stories.tsx
@@ -1,0 +1,9 @@
+import Nav from "../../components/organisms/toolist/nav";
+
+const storyConfig = {
+  title: "Design System/Organisms/Tools Nav"
+};
+
+export default storyConfig;
+
+export const ToolNavOrganism = () => <Nav />;
