+++ stories/organisms/tools-display.stories.tsx
@@ -1,0 +1,9 @@
+import Tool from "../../components/organisms/ToolsDisplay/tools-display";
+
+const storyConfig = {
+  title: "Design System/Organisms/Tools Display"
+};
+
+export default storyConfig;
+
+export const ToolsDisplayOrganism = () => <Tool />;
