+++ stories/organisms/top-nav.stories.tsx
@@ -1,0 +1,9 @@
+import TopNav from "../../components/organisms/TopNav/top-nav";
+
+const storyConfig = {
+  title: "Design System/Organisms/Top Nav"
+};
+
+export default storyConfig;
+
+export const TopNavOrganism = () => <TopNav />;
